renamed weekly input type and remove some logging
