+++ __tests__/weekly/WeeklyInput.js
@@ -1,5 +1,5 @@
 /* eslint-env jest */
-import { getDeliveryDate } from '../../src/weekly/input'
+import { getDeliveryDate } from '../../src/weekly/WeeklyInput'
 import moment from 'moment'
 
 var MockDate = require('mockdate')
